+++ Networking/Sockets/BindingSocket.h
@@ -8,7 +8,7 @@
 namespace pg {
     class BindingSocket: public SimpleSocket{
     public:
-        BindingSocket(int domain, int service, int protocol, int port, uint32_t interface);
+        BindingSocket(int domain, int service, int protocol, short port, uint32_t interface);
         int connectToNetwork(int, struct sockaddr_in);
     };
 }
