+++ Networking/Sockets/ListeningSocket.cpp
@@ -4,7 +4,7 @@
 
 #include "ListeningSocket.h"
 namespace pg    {
-    ListeningSocket::ListeningSocket(int domain, int service, int protocol, int port, uint32_t interface, int bklg) :
+    ListeningSocket::ListeningSocket(int domain, int service, int protocol, short port, uint32_t interface, int bklg) :
     BindingSocket(domain, service, protocol, port, interface)   {
         backlog =bklg;
         startListening();
