+++ Networking/Sockets/ListeningSocket.h
@@ -11,7 +11,7 @@ namespace pg {
         int backlog;
         int listening;
     public:
-        ListeningSocket(int domain, int service, int protocol, int port, uint32_t interface, int bklg);
+        ListeningSocket(int domain, int service, int protocol, short port, uint32_t interface, int bklg);
         void startListening();
     };
 }
