+++ Networking/Sockets/SimpleSocket.h
@@ -16,7 +16,7 @@ namespace pg {
         int sock;
         //int connection;
     public:
-        SimpleSocket(int domain, int service, int protocol, int port, uint32_t interface);
+        SimpleSocket(int domain, int service, int protocol, short port, uint32_t interface);
         virtual int connectToNetwork(int, struct sockaddr_in) =0;
         void testConnection(int );
         struct sockaddr_in getAddress();
